Individual file processing base
