+++ api/functions/main.py
@@ -6,4 +6,5 @@ from sign_files import sign_files
 from storage_watcher import storage_watcher
 
 # Dataset Uploads
-from process_dataset import process_dataset, process_dataset_file
+from process_dataset import process_dataset
+from process_file import process_dataset_file
