+++ architecture/modules/functions/variables.tf
@@ -7,4 +7,6 @@ variable "project" {}
 variable "functions_folder" {}
 variable "elpis_worker" {}
 variable "user_upload_files_bucket" {}
+variable "user_datasets_bucket" {}
+
 variable "dataset_processing_topic" {}
